made minor changes

+++ Cubes/src/cubes/Vector3Float.java
@@ -22,7 +22,7 @@ public class Vector3Float {
 
     public float x, y, z;
 
-    public Vector3Float(int x, int y, int z) {
+    public Vector3Float(float x, float y, float z) {
         this.x = x;
         this.y = y;
         this.z = z;
